Use Genson to create the config file

+++ src/main/java/com/amitg/fantistimeclient/ConfigValues.java
@@ -6,6 +6,6 @@ class ConfigValues {
    public String serverIP;
    public short serverPort;
    public short checkInterval;
-   public short idleSensitivity;
+   public float idleSensitivity;
    public short checksPerReport;
 }
